Put strawberry tasks in a separate array. Removed unnecessary task IDs and weights.
Universal strawberry-based tasks, which are excluded from Reflection, are now in a separate "strawberry" array. Tasks where the task ID is not used now do not have a specified ID and should default to 0. Tasks with a weight of 1 now do not have a specified weight and should default to 1.